MA-8720-1 [Android_MX6Q/SL_SD/EVK] Screen Lock: Can't slide to unlock the screen with USB mouse.100%
The last point is on top of screen usually when unlock the keyguard with mouse.
Event ACTION_UP will be ignored by system and endMotionEvent will not be called.
When current bar state is KEYGUARD, we do not need enable below feature:
"single mouse click expands the panel instead of dragging"

Change-Id: I71a1602bc02e2670b375fa5dbbc835cbecc91982
Signed-off-by: zhang sanshan <b51434@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1087
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/2441
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>