MA-8720-2 [Android_MX6Q/SL_SD/EVK] Screen Lock: Can't slide to unlock the screen with USB mouse.100%
Evaluation have a threshold value which judge whether this stroke is human interaction.
The value of evaluation is main according the current touch event.
The factor which effect evaluation include 11 types, like as Acceleration,SpeedAngles.
The bigger value is, the more unlikely human interaction.
Take SpeedAngles as example:
The speed of a good stroke is most often increases. If the value of SpeedAngles is 0
which means the speed of this stroke increses.

It's hard to unlock the screen, need enlarge the threshold to unlock screen.

Change-Id: Ie6d0ef0e44cfe5146189e64c10c4e56e3f102025
Signed-off-by: zhang sanshan <b51434@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1088
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/2442
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>