MA-18218 Fix bootanimation not correct when config SF_PRIMARY_DISPLAY_ORIENTATION
When config SF_PRIMARY_DISPLAY_ORIENTATION to 90 or 270,bootanimation display
area is not correct at the beginning and it will be ok after displayReady in
SystemServer. The viewport/frame is not correct just after surfaceflinger
setupNewDisplayDevice. displayReady will setProjection and make viewport/frame
to correct one. Here call setDisplayProjection in bootanimation to avoid this
issue.

Change-Id: I254fb0c519a559b6b660dbfa9e8064a559e1d596
Signed-off-by: Jessie Hao <juan.hao@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13939
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17010
Reviewed-on: http://androidsource.ap.freescale.net/project/17206