MA-19123 Fix GTS case testCodecMethodsAccessible
This reverts commit fa27bca26435ae56831fafa092358e4b26ead990,
"Revert "MA-18527 Revert "Move all BluetoothCodecConfig and
BluetoothCodecStatus APIs moved from"

This gts case failed for Accessing hidden method Landroid/bluetooth/Bluetooth
CodecConfig;->getCodecType()I (blacklist, reflection, denied)

the failure is because we remove the api from hidden api list,
delete the "@UnsupportedAppUsage" of this api in
http://androidsource.ap.freescale.net/project/#/c/platform/frameworks/base/+/14573/.
so cause the gts case check this api access, cause the case failed.

so we need to keep align with google, add the api back into hidden api list.

the original reason to revert this patch is to fix CtsHiddenApiBlacklistTestApiTestCases.
now we have fixed this issue by other patch, so this patch is not needed.

Change-Id: I57bd104cae6deb0dffd1490d0d02831ada7bfd3f
Reviewed-on: http://androidsource.ap.freescale.net/project/16048
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>