MA-21142 Fix MediaFrameworkTest: CameraBinderTest failed
The reason why the test fails is that on the board with
one camera connected, the case will test two cameras by default.
Camera nums used is obtained by the packageManager.getSystemAvailableFeatures,
which was not reasonable.
We will change the nums used to the value obtained by the camera service.

Change-Id: I6901a9916af077c728dbc5ecf3c75b23fa337d76
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/21038
Reviewed-by: fang hui <hui.fang@nxp.com>